+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+# [1.207.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.206.0...v1.207.0) (2025-10-24)
+
+
+### Features
+
+* **SessionCTA:** L3-8027 CTA Buttons Cut Off on Sessions & Viewings Modal for RM ([#750](https://github.com/PhillipsAuctionHouse/seldon/issues/750)) ([7448a01](https://github.com/PhillipsAuctionHouse/seldon/commit/7448a010bd9509e3930d4867ac7d6f67e4b923ad))
+
+# [1.206.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.205.2...v1.206.0) (2025-10-23)
+
+
+### Features
+
+* **wizard:** L3-7123 Create progress component (Part 2)  ([#697](https://github.com/PhillipsAuctionHouse/seldon/issues/697)) ([6cbc1a8](https://github.com/PhillipsAuctionHouse/seldon/commit/6cbc1a8c0ed9ec80a5af286c955bc41fcdd231d7))
+
+## [1.205.2](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.205.1...v1.205.2) (2025-10-22)
+
+
+### Bug Fixes
+
+* **Filter:** L3-8017 Calendar filters - CTAs are being cut off and the filter drawer does not fill the screen on mobile browsers Fix ([#749](https://github.com/PhillipsAuctionHouse/seldon/issues/749)) ([db5a8db](https://github.com/PhillipsAuctionHouse/seldon/commit/db5a8dbc0b0be56a4a32da834cafaf6701196654))
+
+## [1.205.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.205.0...v1.205.1) (2025-10-22)
+
+
+### Bug Fixes
+
+* **Filters:** L3-8017 Calendar filters - CTAs are being cut off and the filter drawer does not fill the screen on mobile browsers ([#748](https://github.com/PhillipsAuctionHouse/seldon/issues/748)) ([3d6af11](https://github.com/PhillipsAuctionHouse/seldon/commit/3d6af116eb9064d70a81c709973986e753ae426d))
+
+# [1.205.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.204.2...v1.205.0) (2025-10-21)
+
+
+### Features
+
+* **NotificationBanner:** L3-7834 Update link styles and notification… ([#745](https://github.com/PhillipsAuctionHouse/seldon/issues/745)) ([5761b21](https://github.com/PhillipsAuctionHouse/seldon/commit/5761b21d4be84ec48e65cc626110ec493da6e234))
+
 ## [1.204.2](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.204.1...v1.204.2) (2025-10-15)